Build page structure in HTML instead of js
Refactored many files to delete the code that added everything
to the DOM, since that's now written in the HTML file.
Made a couple of minor changes to some files to make sure that
necessary functions (event listeners etc.) were still called.
Updated the CSS to reflect new names and structures.
Will probably go on to consolidate things quite a bit more.